+++ pkg/service/keystore/service.go
