+++ pkg/service/keystore/service_test.go
